+++ app/src/main/java/com/lurian/easyrecipe/MainActivity.kt
@@ -4,36 +4,14 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
-import androidx.compose.ui.Modifier
-import androidx.navigation.compose.rememberNavController
-import com.lurian.design_system.components.bottomnavbar.BottomNavBar
-import com.lurian.shared.navigation.AppNavHost
-import com.lurian.shared.navigation.bottomNavItems
-import com.lurian.design_system.components.theme.EasyRecipeTheme
+import com.lurian.shared.AppRootView
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
         setContent {
-            EasyRecipeTheme {
-                val navController = rememberNavController()
-                Scaffold(
-                    bottomBar = {
-                        BottomNavBar(
-                            navController = navController,
-                            items = bottomNavItems
-                        )
-                    },
-                ) { innerPadding ->
-                    AppNavHost(
-                        modifier = Modifier.padding(innerPadding),
-                        navController = navController
-                    )
-                }
-            }
+            AppRootView()
         }
     }
 }
