+++ desktop/src/main/kotlin/com/lurian/desktop/Main.kt
@@ -1,5 +1,15 @@
 package com.lurian.desktop
 
+import androidx.compose.ui.window.singleWindowApplication
+import com.lurian.shared.AppRootView
+import com.lurian.shared.SharedInjection
+import org.koin.core.context.startKoin
+
 fun main() {
-    print("desktop")
+    startKoin {
+        modules(SharedInjection.modules())
+    }
+    singleWindowApplication(title = "EasyRecipe") {
+        AppRootView()
+    }
 }
